requirements.txt
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.1.0